extract NeuronTypeDiscovery from IndexService
